requirements.txt
llama-cpp-python
jinja2